add 'show threads' and 'show profile' commands. These are momstly debugging tools for developers, a bit documented in the header files (utils.h), although more documentation is definitely necessary.
The performance impact is close to zero(*) so there is no
need to compile it conditionally.
(*) not completely true - thread destruction still needs
to search a list _but_ this can be easily optimized if we
end up with hundreds of active threads (in which case, though,
the problem is clearly elsewhere).



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@19544 f38db490-d61c-443f-a65b-d21fe96a405b