Two things: 1. slightly rearrange/simplify the parsing of the argument in sip_register.    This brings in a patch that has been in Mantis (5834)  for ages,    and is the larger part of the commit;
2. implement the "contact" option for peers, similar to the one in users.conf:

   If you put a "contact" option with a non-empty argument (e.g. contact=123)
   in a peer section, asterisk will register with the provider as if you had a     

        register= username:secret@host/contact 

   line in the general section.

The latter is a very small is a new feature so i am not putting it
in the 1.4 branch, although the "contact" option in user.conf is
already in the 1.4 branch and so it wouldn't be too strange to
merge it.

Note that the implementation of "contact" is much simpler than
the one in 5834, and limited to a few lines in build_peer().



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@44566 f38db490-d61c-443f-a65b-d21fe96a405b