Merged revisions 51087 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
................
r51087 | file | 2007-01-16 00:55:23 -0500 (Tue, 16 Jan 2007) | 10 lines

Merged revisions 51085 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r51085 | file | 2007-01-16 00:53:31 -0500 (Tue, 16 Jan 2007) | 2 lines

Add none as a valid callgroup/pickupgroup option. I consider it a bug that it would inherit it all the way down and not have any way to reset it to nothing - so that's why it is in 1.2. (issue #8296 reported by gkloepfer)

........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@51090 f38db490-d61c-443f-a65b-d21fe96a405b