Merged revisions 61427 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
................
r61427 | russell | 2007-04-11 10:09:39 -0500 (Wed, 11 Apr 2007) | 14 lines

Merged revisions 61426 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r61426 | russell | 2007-04-11 10:05:36 -0500 (Wed, 11 Apr 2007) | 6 lines

Fix a bug with switching between host=dynamic and using specific hosts for
peers.  The code would only reset the peer's address when it is dynamic if
it was a new peer structure.  Now, it will also reset the address if it was
already in the peer list, but before the reload, it was not dynamic.
(issue #9515, reported by caio1982, fixed by me)

........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@61428 f38db490-d61c-443f-a65b-d21fe96a405b