Merged revisions 67420 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r67420 | murf | 2007-06-05 12:17:28 -0600 (Tue, 05 Jun 2007) | 1 line

Added code to automatically add a default case to switches that don't have one. In some cases, rather than fall thru, it results in a goto with -1 result, which terminates the extension; a sort of dialplan seqfault, sort of. This was required to fix bug reported in 9881
........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@67423 f38db490-d61c-443f-a65b-d21fe96a405b