Merged revisions 77945 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
................
r77945 | murf | 2007-08-02 12:21:40 -0600 (Thu, 02 Aug 2007) | 9 lines

Merged revisions 77942 via svnmerge from 
https://origsvn.digium.com/svn/asterisk/branches/1.2

........
r77942 | murf | 2007-08-02 11:56:37 -0600 (Thu, 02 Aug 2007) | 1 line

This patch hopefully solves 10141; The user is running with it, and it doesn't appear to harm asterisk's operation, and may prevent a crash. I'll store it in 1.2, as we have shut down support on 1.2, but since I developed the patch before support finished, and it might affect 1.4 and trunk, I'm going ahead with it.
........

................


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@77946 f38db490-d61c-443f-a65b-d21fe96a405b