Fix building res_crypto on systems that init locks with constructors.
The problem was that res_crypto now has a RWLIST named "keys".  The macro
for defining this list defines a function used as a constructor for the list
called "init_keys".  However, there was another function called init_keys in
this module for a CLI command.  The fix is just to prepend the generated
functions with underscores.


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@78138 f38db490-d61c-443f-a65b-d21fe96a405b