Merged revisions 88768 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r88768 | russell | 2007-11-05 15:33:56 -0600 (Mon, 05 Nov 2007) | 8 lines

When traversing the list of channel variables here in transmit_invite(), the 
asterisk channel must be locked, as this data may change at any time.

(I have seen numerous reports of crashes related to the handling of channel
variables.  There are a couple of issues on the bug tracker related to it,
but it has also been noted on IRC and mailing lists.  So, I am finding and
fixing some places where channel variables are handled improperly.) 

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@88769 f38db490-d61c-443f-a65b-d21fe96a405b