Merged revisions 91826 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r91826 | russell | 2007-12-07 15:11:08 -0600 (Fri, 07 Dec 2007) | 6 lines

I love fixing lock related errors in the lock debugging code.  That's about as
ironic as it gets in Asterisk programming land.  Anyway, I spotted this bug while
trying to track down why systems are locking up and acting weird in issue #11080.
The mutex attribute object was marked as static in this function when it should
not have been.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@91827 f38db490-d61c-443f-a65b-d21fe96a405b