Merging the queue-penalty branch. In short, this allows one to dynamically adjust the QUEUE_MAX_PENALTY and the newly introduced QUEUE_MIN_PENALTY during a call depending on the amount of time passed. The purpose is to allow the call to open up to more (or maybe just different) members without the caller's losing his place in the queue. See configs/queuerules.conf.sample for an example of how to set up queue rules and configs/queues.conf.sample for how to associate a rule with a queue.
Along with the functional changes, new CLI and manager commands exist to show the rules defined and
there is an additional CLI command to reload the queue rules.

Future enhancements that may be made: support for realtime queue rules and support for dynamically adding
a rule through the manager or CLI. Also a manager command to reload the queue rules (I'll probably write
this myself very soon).



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@94370 f38db490-d61c-443f-a65b-d21fe96a405b