Some changes to app_amd.
The channel name is printed in verbose messages
maximumWordLength option added.
Duration of words that do not meet the minimum word duration will be logged
The duration of pre-greeting silence will be logged
Only consider us in the greeting if we actually detected a valid word duration.

(closes issue #11650, reported and patched by davevg)



git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@95167 f38db490-d61c-443f-a65b-d21fe96a405b