Blocked revisions 104026 via svnmerge
........
r104026 | mmichelson | 2008-02-21 14:12:38 -0600 (Thu, 21 Feb 2008) | 7 lines

Remove an incorrect debug message. It reported that it had received a specific event and tried to report
which event was received. What actually was happening was that it was reporting the number of bytes returned
from a call to read().

Thanks to Jared Smith for bringing the issue up on IRC


........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@104028 f38db490-d61c-443f-a65b-d21fe96a405b