Merged revisions 110035 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r110035 | file | 2008-03-19 16:11:33 -0300 (Wed, 19 Mar 2008) | 4 lines

Add sanity checking for position resuming. We *have* to make sure that the position does not exceed the total number of files present, and we have to make sure that the position's filename is the same as previous. These values can change if a music class is reloaded and give unpredictable behavior.
(closes issue #11663)
Reported by: junky

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@110036 f38db490-d61c-443f-a65b-d21fe96a405b