Merged revisions 114695 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r114695 | seanbright | 2008-04-26 21:26:15 -0400 (Sat, 26 Apr 2008) | 5 lines

When we don't explicitly pass a path to the --with-tds configure option, we
may end up finding tds.h in /usr/local/include instead of /usr/include.  If
this happens, the grep that looks for the version (from tdsver.h) will fail
and we'll have some problems during the build.

........


git-svn-id: http://svn.asterisk.org/svn/asterisk/trunk@114696 f38db490-d61c-443f-a65b-d21fe96a405b